Rename components and update references to better reflect component purpose
