+++ packages/react-learn-card/src/components/SmallAchievementCard/index.ts
@@ -2,4 +2,4 @@
  * @file Automatically generated by barrelsby.
  */
 
-export * from "./SmallVCCard";
+export * from "./SmallAchievementCard";
