+++ packages/react-learn-card/src/components/index.ts
@@ -12,7 +12,7 @@ export * from "./MiniVCThumbnail/index";
 export * from "./Notification/index";
 export * from "./QRCodeCard/index";
 export * from "./RoundedSquare/index";
-export * from "./SmallVCCard/index";
+export * from "./SmallAchievementCard/index";
 export * from "./VCCard/index";
 export * from "./VCDisplayBackFace/index";
 export * from "./VCDisplayCard/index";
